+++ fiddles/node/fiddle-0022-Gulp/app/js/secondary.js
@@ -1,0 +1,1 @@
+console.log("I am logging from the secondary js file.");