+++ fiddles/three/fiddle-0014-EniPuzzle/src/namespaces.js
@@ -1,7 +1,0 @@
-
-app.toolkit = app.toolkit || {};
-app.toolkit.three = app.toolkit.three || {};
-app.view = app.view || {};
-
-
-