+++ fiddles/three/fiddle-0023-EniPuzzle/src/namespaces.js
@@ -1,0 +1,6 @@
+app.toolkit = app.toolkit || {};
+app.toolkit.three = app.toolkit.three || {};
+app.view = app.view || {};
+
+
+