+++ scripts/bin/_env.sh
@@ -13,8 +13,8 @@
 # Baseline Ver - See CHANGELOG @ 201703100420
 # ---------------------------------------------------------------------------------------------------|
 
-HOME_ROOT="/Users/e13542";
+HOME_ROOT="/Users/bradyhouse";
 GITHUB_PUBLISH_REPO="bradyhouse.github.io";
 GITHUB_ROOT="${HOME_ROOT}/github";
-BUILD_NUM="201703100420";
+BUILD_NUM="201704170420";